+++ tests/test_outlierdetection.py
@@ -237,27 +237,33 @@ class TestOutlierDetection(unittest.TestCase):
         self.assertEqual(gooddata_stats.loc['max']['flag'], 0)
         self.assertEqual(gooddata_stats.loc['count']['s_noise'], 1261)
 
-    def test_hampel_filter(self):
+    def test_hampel_filter_basic(self):
         df = ed.load_exampledata_parquet()
         s = df['Tair_f'].copy()
         s = s.loc[s.index.year == 2018].copy()
         s = s.loc[s.index.month == 7].copy()
         s_noise = add_impulse_noise(series=s,
-                                    factor_low=-14,
-                                    factor_high=19,
-                                    contamination=0.06,
+                                    factor_low=-12,
+                                    factor_high=17,
+                                    contamination=0.07,
                                     seed=42)  # Add impulse noise (spikes)
 
         # Checks on noise data, make sure we have outliers, i.e., greater or less than the specified limits
         self.assertGreater(s_noise.max(), 22)
         self.assertLess(s_noise.min(), 10)
 
-        ham = Hampel(
+        ham = HampelDaytimeNighttime(
             series=s_noise,
-            n_sigma=4,
-            window_length=48 * 9,
+            n_sigma_dt=5.5,
+            n_sigma_nt=5.5,
+            window_length=48,
+            use_differencing=False,
+            separate_day_night=False,
             showplot=False,
-            verbose=False
+            verbose=False,
+            lat=47.286417,
+            lon=7.733750,
+            utc_offset=1
         )
         ham.calc(repeat=True)
         flag = ham.get_flag()
@@ -266,20 +272,20 @@ class TestOutlierDetection(unittest.TestCase):
 
         # Checks on bad data
         baddata_stats = checkdf.loc[checkdf.flag == 2].describe()
-        self.assertEqual(baddata_stats.loc['max']['s_noise'], 457.7993015844542)
-        self.assertEqual(baddata_stats.loc['min']['s_noise'], -47.71027579882934)
-        self.assertEqual(baddata_stats.loc['count']['flag'], 82)
+        self.assertEqual(baddata_stats.loc['max']['s_noise'], 420.37816376334473)
+        self.assertEqual(baddata_stats.loc['min']['s_noise'], -38.04507418841196)
+        self.assertEqual(baddata_stats.loc['count']['flag'], 92)
         self.assertEqual(baddata_stats.loc['min']['flag'], 2)
         self.assertEqual(baddata_stats.loc['max']['flag'], 2)
-        self.assertEqual(baddata_stats.loc['count']['s_noise'], 82)
+        self.assertEqual(baddata_stats.loc['count']['s_noise'], 92)
 
         # Checks on good data
         gooddata_stats = checkdf.loc[checkdf.flag == 0].describe()
-        self.assertEqual(gooddata_stats.loc['max']['s_noise'], 25.201551539241972)
-        self.assertEqual(gooddata_stats.loc['min']['s_noise'], 3.65795487099882)
+        self.assertEqual(gooddata_stats.loc['max']['s_noise'], 51.98160999845608)
+        self.assertEqual(gooddata_stats.loc['min']['s_noise'], -7.383053125047635)
         self.assertEqual(gooddata_stats.loc['min']['flag'], 0)
         self.assertEqual(gooddata_stats.loc['max']['flag'], 0)
-        self.assertEqual(gooddata_stats.loc['count']['s_noise'], 1406)
+        self.assertEqual(gooddata_stats.loc['count']['s_noise'], 1396)
 
     def test_zscore(self):
         df = ed.load_exampledata_parquet()